osint_omega/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osint-omega"
version = "0.1.0"
description = "OSINT OMEGA — orchestrateur OSINT à usage légal et éthique"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.9",
    "PyYAML>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.6",
]

[project.scripts]
osint-omega = "osint_omega.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["osint_omega*"]

[tool.setuptools.package-data]
osint_omega = ["agents/*.xml", "agents/*.md"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
